관리자 카운트 매퍼 다오 서비스 수정

+++ src/main/java/com/bugjeogbugjeog/app/bugjeogbugjeog/mapper/BusinessMapper.java
@@ -33,7 +33,7 @@ public interface BusinessMapper {
     public List<BusinessDTO> adminSelectAllBusiness(Criteria criteria);
 
     /* 관리자 카운트 */
-    public int count();
+    public Long count();
 
     /* 관리자 유통 상세 보기 */
     public BusinessDTO adminSelectBusiness(Long businessId);
